fix(instant): update buy quote at start up in the case of default amount
